+++ test/CMakeLists.txt
@@ -47,7 +47,8 @@ primitiv_test(graph)
 primitiv_test(initializer_impl)
 primitiv_test(mixins)
 primitiv_test(model)
-primitiv_test(msgpack)
+primitiv_test(msgpack_objects)
+primitiv_test(msgpack_writer)
 primitiv_test(naive_device)
 primitiv_test(node)
 primitiv_test(parameter)
