+++ test/msgpack_writer_test.cc
@@ -7,7 +7,7 @@
 #include <unordered_map>
 #include <vector>
 #include <gtest/gtest.h>
-#include <primitiv/msgpack.h>
+#include <primitiv/msgpack/writer.h>
 #include <test_utils.h>
 
 using std::string;
